DataViews: Move the layouts into a dedicated folder
