+++ packages/dataviews/src/layouts/index.ts
@@ -12,11 +12,11 @@ import {
 /**
  * Internal dependencies
  */
-import ViewTable from './view-table';
-import ViewGrid from './view-grid';
-import ViewList from './view-list';
-import { LAYOUT_GRID, LAYOUT_LIST, LAYOUT_TABLE } from './constants';
-import type { View } from './types';
+import ViewTable from './table';
+import ViewGrid from './grid';
+import ViewList from './list';
+import { LAYOUT_GRID, LAYOUT_LIST, LAYOUT_TABLE } from '../constants';
+import type { View } from '../types';
 
 export const VIEW_LAYOUTS = [
 	{
